Move asset processing tests to their own file.
